refactor(amplify-gen2-migration-e2e-system): simplify e2e migration system
Remove unused modules and tests from the e2e migration system:
configuration-loader, environment-detector, app-selector,
category-initializer, cdk-atmosphere-integration, directory-manager,
file-manager, aws-profile-manager, types, and math. Streamline the
CLI entry point and core components to reduce complexity.

Restructure project-boards app by moving backend files into
backend/ and migration scripts into migration/ subdirectories.

Fix lock.ts to use discover() for finding the GraphQL API ID
instead of paginating AppSync APIs directly.

Add git utility module, frontest.ts, and cleanup-s3-buckets.sh.
---
Prompt: Commit everything I've done. Dont run tests, don't do
anything. just commit.
