+++ .eslint-dictionary.json
@@ -161,6 +161,7 @@
   "formatter",
   "frontend",
   "frontends",
+  "frontest",
   "fsext",
   "func",
   "funcs",
