--- amplify-migration-apps/project-boards/configure.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-cp -f schema.graphql ./amplify/backend/api/projectboards/schema.graphql
-cp -f quotegenerator.js ./amplify/backend/function/quotegenerator/src/index.js
