--- amplify-migration-apps/project-boards/gen-1-cleanup.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-rm -rf amplify/
-rm src/amplifyconfiguration.json
-rm src/aws-exports.js
-
-exit 0
