+++ amplify-migration-apps/project-boards/package.json
@@ -11,7 +11,7 @@
     "build": "tsc -b && vite build",
     "lint": "eslint .",
     "preview": "vite preview",
-    "configure": "./configure.sh",
+    "configure": "./backend/configure.sh",
     "sanitize": "tsx ../sanitize.ts",
     "typecheck": "cd _snapshot.post.generate/amplify && npx tsc --noEmit"
   },
@@ -23,6 +23,7 @@
     "react-dom": "^19.1.1"
   },
   "devDependencies": {
+    "@aws-sdk/client-cognito-identity-provider": "^3.936.0",
     "@eslint/js": "^9.36.0",
     "@types/node": "^24.6.0",
     "@types/react": "^19.1.16",
