+++ packages/amplify-gen2-migration-e2e-system/src/cli.ts
@@ -1,4 +1,6 @@
 #!/usr/bin/env node
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
 
 /**
  * CLI entry point for the Amplify Migration System
@@ -7,57 +9,20 @@
 // eslint-disable-next-line spellcheck/spell-checker
 import * as yargs from 'yargs';
 import chalk from 'chalk';
-import { Logger } from './utils/logger';
-import { FileManager } from './utils/file-manager';
-import { ConfigurationLoader } from './core/configuration-loader';
-import { EnvironmentDetector } from './core/environment-detector';
-import { AppSelector } from './core/app-selector';
+import { Logger, LogLevel } from './utils/logger';
 import { AmplifyInitializer } from './core/amplify-initializer';
-import { CategoryInitializer } from './core/category-initializer';
-import { DirectoryManager } from './utils/directory-manager';
-import { CDKAtmosphereIntegration } from './core/cdk-atmosphere-integration';
 import { Gen2MigrationExecutor } from './core/gen2-migration-executor';
 import execa from 'execa';
-import { LogLevel, CLIOptions, AppConfiguration, EnvironmentType, InitializeAppFromCLIParams } from './types';
-import { generateTimeBasedE2EAmplifyAppName } from './utils/math';
 import path from 'path';
 import os from 'os';
 import fs from 'fs';
 import * as fsExtra from 'fs-extra';
 import { getCLIPath } from '@aws-amplify/amplify-e2e-core';
-
-/** Options passed to app-specific post-generate scripts */
-interface PostGenerateOptions {
-  appPath: string;
-  envName?: string;
-}
-
-/** Options passed to app-specific post-refactor scripts */
-interface PostRefactorOptions {
-  appPath: string;
-  envName?: string;
-}
-
-/** Shape of an app's post-generate module */
-interface PostGenerateModule {
-  postGenerate: (options: PostGenerateOptions) => Promise<void>;
-}
-
-/** Shape of an app's post-refactor module */
-interface PostRefactorModule {
-  postRefactor: (options: PostRefactorOptions) => Promise<void>;
-}
+import * as git from './utils/git';
 
 // Initialize core components
 const logger = new Logger(LogLevel.INFO);
-const fileManager = new FileManager(logger);
-const configurationLoader = new ConfigurationLoader(logger, fileManager);
-const environmentDetector = new EnvironmentDetector(logger);
-const appSelector = new AppSelector(logger, fileManager);
 const amplifyInitializer = new AmplifyInitializer(logger);
-const categoryInitializer = new CategoryInitializer(logger);
-const directoryManager = new DirectoryManager(logger);
-const cdkAtmosphereIntegration = new CDKAtmosphereIntegration(logger, environmentDetector);
 const gen2MigrationExecutor = new Gen2MigrationExecutor(logger);
 
 // Default migration target directory
@@ -75,12 +40,6 @@ async function main(): Promise<void> {
         description: 'App to migrate (e.g., project-boards)',
         string: true,
       })
-      .option('dry-run', {
-        alias: 'd',
-        type: 'boolean',
-        description: 'Show what would be done without executing',
-        default: false,
-      })
       .option('verbose', {
         alias: 'v',
         type: 'boolean',
@@ -92,28 +51,11 @@ async function main(): Promise<void> {
         description: 'AWS profile to use',
         string: true,
       })
-      .option('atmosphere', {
-        type: 'boolean',
-        description: 'Use atmosphere credentials in execution environment',
-      })
-      .option('envName', {
-        type: 'string',
-        description: 'Amplify env name to create',
-        string: true,
-      })
-      .option('list-apps', {
-        alias: 'l',
-        type: 'boolean',
-        description: 'List available apps and exit',
-        default: false,
-      })
       .help()
       .alias('help', 'h')
       .version()
       .alias('version', 'V')
-      .example('$0 -a project-boards', 'Migrate specific app')
-      .example('$0 --dry-run', 'Show what would be done')
-      .example('$0 --list-apps', 'List all available apps').argv;
+      .example('$0 -a project-boards', 'Migrate specific app').argv;
 
     // Set log level based on verbose flag
     if (argv.verbose) {
@@ -123,79 +65,31 @@ async function main(): Promise<void> {
     // Print banner
     printBanner();
 
-    // Handle special commands
-    if (argv['list-apps']) {
-      await handleListApps();
-      return;
-    }
-
     // Validate required options for migration
     if (!argv.app) {
       logger.error('Error: --app is required for migration');
       process.exit(1);
     }
 
-    if (!argv.profile && !argv.atmosphere) {
-      throw new Error('Either --profile or --atmosphere must be specified');
+    if (!argv.profile) {
+      throw new Error('--profile must be specified');
     }
 
-    // Build CLI options
-    const options: CLIOptions = {
-      app: argv.app,
-      dryRun: argv['dry-run'],
-      verbose: argv.verbose,
-      profile: argv.profile,
-      isAtmosphere: argv.atmosphere,
-      envName: argv.envName,
-    };
-
     // Select apps to process
-    logger.debug('Selecting apps for migration...');
-    await appSelector.validateAppExists(options.app);
-    const selectedApp = options.app;
+    const appPath = path.join(__dirname, '..', '..', '..', 'amplify-migration-apps', argv.app);
+    if (!fsExtra.existsSync(appPath)) {
+      throw new Error(`App not found: ${appPath}`);
+    }
+    const selectedApp = argv.app;
     const deploymentName = generateTimeBasedE2EAmplifyAppName(selectedApp);
 
     logger.setAppName(deploymentName);
 
-    // Detect environment and get credentials if needed
-    logger.debug('Detecting execution environment...');
-    let environment;
-
-    if (argv.profile) {
-      environment = EnvironmentType.LOCAL;
-    } else if (argv.atmosphere) {
-      environment = EnvironmentType.ATMOSPHERE;
-      const didValidateAtmosphereEnvVars = await environmentDetector.isAtmosphereEnvironment();
-      if (!didValidateAtmosphereEnvVars) {
-        throw Error('Atmosphere environment requested but required environment variables are not set');
-      } else {
-        logger.info('Atmosphere environment validated successfully');
-      }
-    }
-
     // Get appropriate credentials based on environment
-    let profile: string;
-    const isAtmosphereEnv = environment === EnvironmentType.ATMOSPHERE;
-
-    if (isAtmosphereEnv) {
-      logger.info('Atmosphere environment detected - obtaining atmosphere credentials...');
-
-      try {
-        profile = await cdkAtmosphereIntegration.getProfileFromAllocation();
-        logger.info(`Successfully created Atmosphere AWS profile: ${profile}`);
-      } catch (atmosphereError) {
-        logger.error(`Failed to get atmosphere credentials: ${(atmosphereError as Error).message}`);
-        logger.error('Cannot proceed without atmosphere credentials in atmosphere environment');
-        throw new Error(`Atmosphere environment detected but credentials unavailable: ${(atmosphereError as Error).message}`);
-      }
-    } else {
-      logger.info(`Local environment detected - will use local AWS profile: ${options.profile} for credentials`);
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      profile = options.profile!;
-    }
+    const profile = argv.profile;
 
     // Generate envName if not provided via CLI
-    const envName = options.envName ?? AmplifyInitializer.generateRandomEnvName();
+    const envName = AmplifyInitializer.generateRandomEnvName();
     logger.info(`Using Amplify environment name: ${envName}`);
 
     // Enable file logging
@@ -204,42 +98,14 @@ async function main(): Promise<void> {
 
     logger.setLogFilePath(logFile);
 
-    // Load configuration for selected app
-    logger.debug('Loading app configuration...');
-
-    let config: AppConfiguration | undefined;
-
-    try {
-      config = await configurationLoader.loadAppConfiguration(selectedApp);
-    } catch (error) {
-      logger.error(`Failed to load configuration for ${selectedApp}`, error as Error);
-      throw error;
-    }
-
-    // Handle dry-run mode
-    if (options.dryRun) {
-      logger.info('Dry run mode - showing what would be done:');
-      await showDryRunSummary(selectedApp, config);
-      return;
-    }
-
     // Initialize app
-    const migrationTargetPath = MIGRATION_TARGET_DIR;
-
-    try {
-      await initializeAppFromCLI({ appName: selectedApp, deploymentName, config, migrationTargetPath, envName, profile });
-    } finally {
-      // Cleanup atmosphere profile if we created one
-      if (isAtmosphereEnv) {
-        logger.info('Cleaning up atmosphere resources...');
-        try {
-          await cdkAtmosphereIntegration.cleanup();
-          logger.info('Successfully cleaned up atmosphere resources');
-        } catch (cleanupError) {
-          logger.warn(`Failed to cleanup atmosphere resources: ${(cleanupError as Error).message}`);
-        }
-      }
-    }
+    await initializeAppFromCLI({
+      appName: selectedApp,
+      sourceAppPath: appPath,
+      deploymentName,
+      envName,
+      profile,
+    });
   } catch (error) {
     logger.error('Migration failed', error as Error);
     process.exit(1);
@@ -249,146 +115,58 @@ async function main(): Promise<void> {
 function printBanner(): void {
   console.log(
     chalk.cyan(`
-╔══════════════════════════════════════════════════════════════╗
-║                                                              ║
-║           AWS Amplify Gen1 to Gen2 Migration System          ║
-║                                                              ║
-║  Automation for migrating Amplify applications from          ║
-║           Gen1 to Gen2                                       ║
-║                                                              ║
-╚══════════════════════════════════════════════════════════════╝
+╔══════════════════════════════════════════════════════════════════════╗
+║                                                                      ║
+║           AWS Amplify Gen1 to Gen2 Migration E2E                     ║
+║                                                                      ║
+║  Automation for migrating Amplify applications from Gen1 to Gen2     ║
+║                                                                      ║
+╚══════════════════════════════════════════════════════════════════════╝
 `),
   );
 }
 
-async function handleListApps(): Promise<void> {
-  logger.info('Listing available apps...');
-
-  try {
-    const availableApps = await appSelector.discoverAvailableApps();
-
-    if (availableApps.length === 0) {
-      console.log(chalk.yellow('No apps found in the apps directory'));
-      return;
-    }
-
-    console.log(chalk.green(`\nFound ${availableApps.length} available apps:\n`));
-
-    for (const appName of availableApps) {
-      const appPath = appSelector.getAppPath(appName);
-      console.log(`  ${chalk.cyan(appName.padEnd(10))} ${appPath}`);
-    }
-  } catch (error) {
-    logger.error('Failed to list apps', error as Error);
-    process.exit(1);
-  }
-}
-
-async function showDryRunSummary(selectedApp: string, config?: AppConfiguration): Promise<void> {
-  console.log(chalk.yellow('\n=== DRY RUN SUMMARY ===\n'));
-  console.log('');
-
-  console.log(chalk.cyan('Initialization Actions:'));
-  console.log(`  Migration target directory: ${MIGRATION_TARGET_DIR}`);
-  console.log('');
-  console.log(chalk.cyan(`${selectedApp}:`));
-
-  if (config) {
-    const appConfig = config as AppConfiguration;
-    const categories = Object.keys(appConfig.categories || {});
-    const sourceDir = appSelector.getAppPath(selectedApp);
-
-    console.log(`  Config app name: ${appConfig.app.name}`);
-    console.log(`  Source directory: ${sourceDir}`);
-    console.log(`  Would copy to: ${MIGRATION_TARGET_DIR}/<generated-deployment-name>`);
-    console.log(`  Would run: amplify init with generated deployment name`);
-    console.log(`  Categories: ${categories.join(', ') || 'None'}`);
-  } else {
-    console.log(chalk.red('  Configuration not loaded'));
-  }
-  console.log('');
-
-  console.log(chalk.yellow('=== END DRY RUN SUMMARY ===\n'));
-}
-
 /**
- * Run the app-specific post-generate script if it exists.
- * Each app in amplify-migration-apps can have a post-generate.ts that applies
- * manual edits required after `amplify gen2-migration generate`.
+ * Generates a time-based Amplify app name with optional app name suffix.
+ * Format: [last8alphanumeric][YYMMDDHHMMSS] (20 chars for Amplify compatibility)
+ * CDK resource names (based off of amplify app name) must start with an alphabetic character.
+ * @param appName Optional app name from which to extract last 8 alphanumeric characters
+ * @returns A unique, sortable app name starting with a letter (max 20 chars)
  */
-async function runPostGenerateScript(appName: string, targetAppPath: string, envName?: string): Promise<void> {
-  const sourceAppPath = appSelector.getAppPath(appName);
-  const postGeneratePath = path.join(sourceAppPath, 'post-generate.ts');
+export const generateTimeBasedE2EAmplifyAppName = (appName: string): string => {
+  const now = new Date();
 
-  if (!fs.existsSync(postGeneratePath)) {
-    logger.debug(`No post-generate script found for ${appName} at ${postGeneratePath}`);
-    return;
-  }
+  // Format: YYMMDDHHMMSSMM (human-readable, sortable) - 14 chars
+  const timestamp = [
+    String(now.getFullYear()).slice(-2), // YY
+    String(now.getMonth() + 1).padStart(2, '0'), // MM
+    String(now.getDate()).padStart(2, '0'), // DD
+    String(now.getHours()).padStart(2, '0'), // HH
+    String(now.getMinutes()).padStart(2, '0'), // MM
+  ].join('');
 
-  logger.info(`Running post-generate script for ${appName}...`);
+  // Extract last 8 alphanumeric characters from appName if provided
+  // Total: 10 (prefix) + 10 (timestamp) = 20 chars (at 20 char limit)
+  const alphanumericOnly = appName.replace(/[^a-zA-Z0-9]/g, '');
+  const prefix = alphanumericOnly.slice(0, 10).toLowerCase();
+  // Ensure prefix starts with a letter to avoid CDK resource naming issues
+  const safePrefix = /^[a-z]/.test(prefix) ? prefix : `e${prefix.slice(1)}`;
+  return `${safePrefix}${timestamp}`;
+};
 
-  const postGenerateModule = (await import(postGeneratePath)) as PostGenerateModule;
-
-  if (typeof postGenerateModule.postGenerate !== 'function') {
-    throw new Error(`post-generate.ts for ${appName} does not export a postGenerate function`);
-  }
-
-  await postGenerateModule.postGenerate({ appPath: targetAppPath, envName });
-  logger.info(`Post-generate script completed for ${appName}`);
+async function runGen1TestScript(targetAppPath: string): Promise<void> {
+  await runAppScript(targetAppPath, 'frontest.ts', [path.join('src', 'amplifyconfiguration.json')]);
 }
 
-/**
- * Run the app-specific post-refactor script if it exists.
- * Each app in amplify-migration-apps can have a post-refactor.ts that applies
- * manual edits required after `amplify gen2-migration refactor`.
- */
-async function runPostRefactorScript(appName: string, targetAppPath: string, envName?: string): Promise<void> {
-  const sourceAppPath = appSelector.getAppPath(appName);
-  const postRefactorPath = path.join(sourceAppPath, 'post-refactor.ts');
-
-  if (!fs.existsSync(postRefactorPath)) {
-    logger.debug(`No post-refactor script found for ${appName} at ${postRefactorPath}`);
-    return;
-  }
-
-  logger.info(`Running post-refactor script for ${appName}...`);
-
-  const postRefactorModule = (await import(postRefactorPath)) as PostRefactorModule;
-
-  if (typeof postRefactorModule.postRefactor !== 'function') {
-    throw new Error(`post-refactor.ts for ${appName} does not export a postRefactor function`);
-  }
-
-  await postRefactorModule.postRefactor({ appPath: targetAppPath, envName });
-  logger.info(`Post-refactor script completed for ${appName}`);
+async function runGen2TestScript(targetAppPath: string): Promise<void> {
+  await runAppScript(targetAppPath, 'frontest.ts', ['amplify_outputs.json']);
 }
 
-/**
- * Run the app's gen1-test-script.ts to validate the Gen1 deployment.
- *
- * Copies _test-common to the migration target directory so relative
- * imports like ../_test-common resolve, then executes the test script
- * via npx tsx from the target app directory.
- */
-async function runGen1TestScript(targetAppPath: string, migrationTargetPath: string, sourceAppsBasePath: string): Promise<void> {
-  const testScriptName = 'gen1-test-script.ts';
+async function runAppScript(targetAppPath: string, scriptPath: string, args: string[]): Promise<void> {
+  const testScriptPath = path.join(targetAppPath, scriptPath);
 
-  // Copy _test-common so ../_test-common imports resolve from the target app dir
-  const testCommonSource = path.join(sourceAppsBasePath, '_test-common');
-  const testCommonDest = path.join(migrationTargetPath, '_test-common');
-  logger.info(`Copying _test-common to ${testCommonDest}`);
-  await fsExtra.copy(testCommonSource, testCommonDest, { overwrite: true });
-
-  // Install dependencies for the test script (aws-amplify, etc.)
-  logger.info(`Installing dependencies in ${targetAppPath}`);
-  await execa('npm', ['install'], { cwd: targetAppPath });
-
-  // Install dependencies for _test-common
-  logger.info(`Installing _test-common dependencies in ${testCommonDest}`);
-  await execa('npm', ['install'], { cwd: testCommonDest });
-
-  logger.info(`Running ${testScriptName} in ${targetAppPath}`);
-  const result = await execa('npx', ['tsx', testScriptName], {
+  logger.info(`Running ${scriptPath} in ${targetAppPath} with args: ${args.join(' ')}`);
+  const result = await execa('npx', ['tsx', scriptPath, ...args], {
     cwd: targetAppPath,
     stdio: 'inherit',
     reject: false,
@@ -396,49 +174,8 @@ async function runGen1TestScript(targetAppPath: string, migrationTargetPath: str
   });
 
   if (result.exitCode !== 0) {
-    throw new Error(`${testScriptName} failed with exit code ${result.exitCode}\n`);
+    throw new Error(`${testScriptPath} failed with exit code ${result.exitCode}`);
   }
-
-  logger.info(`${testScriptName} completed successfully`);
-}
-
-/**
- * Run the app's gen2-test-script.ts to validate the Gen2 deployment.
- *
- * Same pattern as runGen1TestScript but runs gen2-test-script.ts instead.
- * Copies _test-common and installs deps before executing.
- */
-async function runGen2TestScript(targetAppPath: string, migrationTargetPath: string, sourceAppsBasePath: string): Promise<void> {
-  const testScriptName = 'gen2-test-script.ts';
-
-  // Check if the gen2 test script exists
-  if (!fs.existsSync(path.join(targetAppPath, testScriptName))) {
-    logger.debug(`No ${testScriptName} found in ${targetAppPath}, skipping`);
-    return;
-  }
-
-  // Copy _test-common so ../_test-common imports resolve from the target app dir
-  const testCommonSource = path.join(sourceAppsBasePath, '_test-common');
-  const testCommonDest = path.join(migrationTargetPath, '_test-common');
-  logger.info(`Copying _test-common to ${testCommonDest}`);
-  await fsExtra.copy(testCommonSource, testCommonDest, { overwrite: true });
-
-  // Install dependencies for _test-common
-  logger.info(`Installing _test-common dependencies in ${testCommonDest}`);
-  await execa('npm', ['install'], { cwd: testCommonDest });
-
-  logger.info(`Running ${testScriptName} in ${targetAppPath}`);
-  const result = await execa('npx', ['tsx', testScriptName], {
-    cwd: targetAppPath,
-    stdio: 'inherit',
-    reject: false,
-  });
-
-  if (result.exitCode !== 0) {
-    throw new Error(`${testScriptName} failed with exit code ${result.exitCode}\n`);
-  }
-
-  logger.info(`${testScriptName} completed successfully`);
 }
 
 /**
@@ -449,14 +186,13 @@ async function runGen2TestScript(targetAppPath: string, migrationTargetPath: str
  */
 async function amplifyPush(targetAppPath: string): Promise<void> {
   const amplifyPath = getCLIPath(true);
-  logger.info(`Using amplify CLI at: ${amplifyPath}`);
+  logger.debug(`Using amplify CLI at: ${amplifyPath}`);
   const originalCwd = process.cwd();
 
   process.chdir(targetAppPath);
   try {
     const result = await execa(amplifyPath, ['push', '--yes', '--debug'], {
       cwd: targetAppPath,
-      stdio: logger.isDebug() ? 'inherit' : 'pipe',
     });
 
     if (result.exitCode !== 0) {
@@ -467,6 +203,34 @@ async function amplifyPush(targetAppPath: string): Promise<void> {
   }
 }
 
+async function amplifyStatus(targetAppPath: string): Promise<void> {
+  const amplifyPath = getCLIPath(true);
+  logger.debug(`Using amplify CLI at: ${amplifyPath}`);
+  const originalCwd = process.cwd();
+
+  process.chdir(targetAppPath);
+  try {
+    const result = await execa(amplifyPath, ['status'], {
+      cwd: targetAppPath,
+      stdio: 'inherit',
+    });
+
+    if (result.exitCode !== 0) {
+      throw new Error(`amplify status failed with exit code ${result.exitCode}`);
+    }
+  } finally {
+    process.chdir(originalCwd);
+  }
+}
+
+export interface InitializeAppFromCLIParams {
+  appName: string;
+  sourceAppPath: string;
+  deploymentName: string;
+  envName: string;
+  profile: string;
+}
+
 /**
  * Initialize a single app.
  * Copies the source directory to the migration target, runs amplify init,
@@ -474,25 +238,20 @@ async function amplifyPush(targetAppPath: string): Promise<void> {
  * full gen2-migration workflow.
  */
 async function initializeAppFromCLI(params: InitializeAppFromCLIParams): Promise<void> {
-  const { appName, deploymentName, config, migrationTargetPath, envName, profile } = params;
+  const { appName, sourceAppPath, deploymentName, envName, profile } = params;
   logger.info(`Starting initialization for ${appName} with deployment name: ${deploymentName}`);
 
-  const sourceAppPath = appSelector.getAppPath(appName);
   logger.debug(`Source app path: ${sourceAppPath}`);
-  logger.debug(`Config app name: ${config.app.name}`);
-
-  // Derive sourceAppsBasePath once for all test script calls
-  const sourceAppsBasePath = path.dirname(sourceAppPath);
 
   try {
     // Create target directory and copy source
-    const targetAppPath = await directoryManager.createAppDirectory({
-      basePath: migrationTargetPath,
-      appName: deploymentName,
-    });
-
+    const targetAppPath = path.join(MIGRATION_TARGET_DIR, deploymentName);
     logger.info(`Copying source directory to target...`);
-    await directoryManager.copyDirectory(sourceAppPath, targetAppPath);
+    fsExtra.mkdirSync(targetAppPath);
+    fsExtra.copySync(sourceAppPath, targetAppPath, {
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars
+      filter: (src: string, _dst: string) => !src.includes('_snapshot') && !src.includes('node_modules'),
+    });
 
     // Update package.json name to use deploymentName for predictable Gen2 stack naming
     const packageJsonPath = path.join(targetAppPath, 'package.json');
@@ -507,133 +266,88 @@ async function initializeAppFromCLI(params: InitializeAppFromCLIParams): Promise
     logger.debug(`Using AWS profile '${profile}' for Amplify initialization`);
     await amplifyInitializer.initializeApp({
       appPath: targetAppPath,
-      config,
       deploymentName,
       envName,
       profile,
     });
 
-    // Initialize categories (auth, api, storage, function, etc.)
-    logger.info(`Initializing categories for ${deploymentName}...`);
-    const categoryResult = await categoryInitializer.initializeCategories({
-      appPath: targetAppPath,
-      config,
-      deploymentName,
-    });
+    logger.info('Configuring categories');
+    await configure(sourceAppPath, targetAppPath);
+    logger.info('Finished configuring categories');
 
-    if (categoryResult.initializedCategories.length > 0) {
-      logger.info(`Successfully initialized categories: ${categoryResult.initializedCategories.join(', ')}`);
-    }
-    if (categoryResult.skippedCategories.length > 0) {
-      logger.warn(`Skipped categories: ${categoryResult.skippedCategories.join(', ')}`);
-    }
-    if (categoryResult.errors.length > 0) {
-      for (const error of categoryResult.errors) {
-        logger.error(`Category '${error.category}' failed: ${error.error}`, undefined);
-      }
-      throw new Error(`Failed to initialize ${categoryResult.errors.length} category(ies)`);
-    }
+    await git.init(targetAppPath);
+    await git.commit(targetAppPath, 'feat: gen1 initial commit');
 
-    // Run configure.sh if present (copies custom source files into amplify/backend/)
-    const configureScriptPath = path.join(targetAppPath, 'configure.sh');
-    if (fs.existsSync(configureScriptPath)) {
-      logger.info(`Running configure.sh for ${deploymentName}...`);
-      await execa('bash', ['configure.sh'], { cwd: targetAppPath, stdio: 'inherit' });
-      logger.info(`Successfully ran configure.sh for ${deploymentName}`);
-    } else {
-      logger.debug(`No configure.sh found for ${deploymentName}, skipping`);
-    }
+    logger.info('Installing dependencies...');
+    await execa('npm', ['install'], { cwd: targetAppPath });
+    logger.info('Finished installing dependencies');
 
-    // Push the initialized app to AWS
+    await amplifyStatus(targetAppPath);
+
     logger.info(`Pushing ${deploymentName} to AWS...`);
     await amplifyPush(targetAppPath);
     logger.info(`Successfully pushed ${deploymentName} to AWS`);
 
-    // Run gen1 test script to validate the Gen1 deployment
-    logger.info(`Running gen1 test script (post-push) for ${deploymentName}...`);
-    await runGen1TestScript(targetAppPath, migrationTargetPath, sourceAppsBasePath);
-    logger.info(`Gen1 test script passed (post-push) for ${deploymentName}`);
+    await git.commit(targetAppPath, 'feat: post push');
 
-    // Initialize git repo and commit the Gen1 state
-    logger.info(`Initializing git repository for ${deploymentName}...`);
-    await execa('git', ['init'], { cwd: targetAppPath });
-    await execa('git', ['add', '.'], { cwd: targetAppPath });
-    await execa('git', ['commit', '-m', 'feat: gen1 initial commit'], { cwd: targetAppPath });
-    logger.info(`Git repository initialized and Gen1 state committed`);
+    await runGen1TestScript(targetAppPath);
 
-    // Run gen2-migration pre-deployment workflow (lock -> checkout -> generate)
-    logger.info(`Running gen2-migration pre-deployment workflow for ${deploymentName}...`);
-    await gen2MigrationExecutor.runPreDeploymentWorkflow(targetAppPath, envName);
-    logger.info(`Successfully completed gen2-migration pre-deployment workflow for ${deploymentName}`);
-
-    // Run app-specific post-generate script
-    await runPostGenerateScript(appName, targetAppPath, envName);
-
-    // Commit Gen2 generated code
-    logger.info(`Committing Gen2 generated code for ${deploymentName}...`);
-    await execa('git', ['add', '.'], { cwd: targetAppPath });
-    await execa('git', ['commit', '-m', 'feat: gen2 migration generate'], { cwd: targetAppPath });
-    logger.info(`Gen2 generated code committed`);
-
-    // Deploy Gen2 using ampx sandbox
-    logger.info(`Deploying Gen2 app using ampx sandbox for ${deploymentName}...`);
     const gen2BranchName = `gen2-${envName}`;
+
+    await gen2MigrationExecutor.assess(targetAppPath);
+    await gen2MigrationExecutor.lock(targetAppPath);
+    await git.checkout(targetAppPath, gen2BranchName, true);
+
+    await gen2MigrationExecutor.generate(targetAppPath);
+    await runAppScript(targetAppPath, path.join('migration', 'post-generate.ts'), [targetAppPath]);
+    await git.commit(targetAppPath, 'feat: gen2 migration generate');
+
     const gen2StackName = await gen2MigrationExecutor.deployGen2Sandbox(targetAppPath, deploymentName, gen2BranchName);
-    logger.info(`Gen2 app deployed with stack name: ${gen2StackName}`);
+    await runGen1TestScript(targetAppPath);
+    await runGen2TestScript(targetAppPath);
 
-    // Run gen2 test script to validate the Gen2 code before deploying
-    logger.info(`Running gen2 test script (post-generate) for ${deploymentName}...`);
-    await runGen2TestScript(targetAppPath, migrationTargetPath, sourceAppsBasePath);
-    logger.info(`Gen2 test script passed (post-generate) for ${deploymentName}`);
-
-    // Checkout back to main branch for refactor (refactor must run from Gen1 branch)
-    logger.info(`Checking out main branch for refactor (refactor requires Gen1 files)...`);
-    await execa('git', ['add', '.'], { cwd: targetAppPath });
-    await execa('git', ['commit', '--allow-empty', '-m', 'chore: before refactor'], { cwd: targetAppPath });
-    await execa('git', ['checkout', 'main'], { cwd: targetAppPath });
-    logger.info('Installing dependencies...');
-    await execa('npm', ['install'], { cwd: targetAppPath });
-
-    // Run refactor to move stateful resources from Gen1 to Gen2
-    logger.info(`Running gen2-migration refactor for ${deploymentName}...`);
+    await git.checkout(targetAppPath, 'main', false);
     await gen2MigrationExecutor.refactor(targetAppPath, gen2StackName);
-    logger.info(`Successfully completed gen2-migration refactor for ${deploymentName}`);
+    await runGen1TestScript(targetAppPath);
+    await runGen2TestScript(targetAppPath);
 
-    // Run gen1 test script to validate post-refactor state
-    logger.info(`Running gen1 test script (post-refactor) for ${deploymentName}...`);
-    await runGen1TestScript(targetAppPath, migrationTargetPath, sourceAppsBasePath);
-    logger.info(`Gen1 test script passed (post-refactor) for ${deploymentName}`);
+    await git.checkout(targetAppPath, gen2BranchName, false);
+    await runAppScript(targetAppPath, path.join('migration', 'post-refactor.ts'), [targetAppPath]);
+    await git.commit(targetAppPath, 'fix: post-refactor edits');
 
-    // Checkout back to Gen2 branch for post-refactor edits
-    logger.info(`Checking out ${gen2BranchName} branch for post-refactor edits...`);
-    await execa('git', ['add', '.'], { cwd: targetAppPath });
-    await execa('git', ['commit', '--allow-empty', '-m', 'chore: after refactor'], { cwd: targetAppPath });
-    await execa('git', ['checkout', gen2BranchName], { cwd: targetAppPath });
-
-    // Run app-specific post-refactor script
-    await runPostRefactorScript(appName, targetAppPath, envName);
-
-    // Commit post-refactor changes
-    logger.info(`Committing post-refactor changes for ${deploymentName}...`);
-    await execa('git', ['add', '.'], { cwd: targetAppPath });
-    await execa('git', ['commit', '-m', 'fix: post-refactor edits'], { cwd: targetAppPath });
-    logger.info(`Post-refactor changes committed`);
-
-    // Redeploy Gen2 to pick up post-refactor changes
-    logger.info(`Redeploying Gen2 app after refactor for ${deploymentName}...`);
     await gen2MigrationExecutor.deployGen2Sandbox(targetAppPath, deploymentName, gen2BranchName);
-    logger.info(`Gen2 app redeployed successfully`);
-
-    // Run gen1 test script to validate final deployment
-    logger.info(`Running gen1 test script (post-redeployment) for ${deploymentName}...`);
-    await runGen1TestScript(targetAppPath, migrationTargetPath, sourceAppsBasePath);
-    logger.info(`Gen1 test script passed (post-redeployment) for ${deploymentName}`);
+    await runGen1TestScript(targetAppPath);
+    await runGen2TestScript(targetAppPath);
 
     logger.info(`App ${deploymentName} fully initialized and migrated at ${targetAppPath}`);
   } catch (error) {
     logger.error(`Failed to initialize ${appName}`, error as Error);
     throw error;
   }
+}
+
+async function configure(sourceAppPath: string, targetAppPath: string): Promise<void> {
+  const restore = (p: string): void => {
+    fsExtra.removeSync(path.join(targetAppPath, 'amplify', p));
+    fsExtra.copySync(path.join(targetAppPath, '.amplify.init', p), path.join(targetAppPath, 'amplify', p));
+  };
+
+  const metaPath = path.join(targetAppPath, 'amplify', 'backend', 'amplify-meta.json');
+  const oldMeta = JSON.parse(fsExtra.readFileSync(metaPath, { encoding: 'utf-8' }));
+
+  fsExtra.moveSync(path.join(targetAppPath, 'amplify'), path.join(targetAppPath, '.amplify.init'));
+  fsExtra.copySync(path.join(sourceAppPath, '_snapshot.pre.generate', 'amplify'), path.join(targetAppPath, 'amplify'));
+
+  restore(path.join('#current-cloud-backend'));
+  restore(path.join('.config'));
+  restore(path.join('team-provider-info.json'));
+
+  const newMeta = JSON.parse(fsExtra.readFileSync(metaPath, { encoding: 'utf-8' }));
+
+  newMeta.providers['awscloudformation'] = oldMeta.providers['awscloudformation'];
+
+  fsExtra.writeFileSync(metaPath, JSON.stringify(newMeta, null, 2));
+  fsExtra.removeSync(path.join(targetAppPath, '.amplify.init'));
 }
 
 process.on('uncaughtException', (error) => {
