+++ packages/amplify-gen2-migration-e2e-system/src/core/amplify-initializer.ts
@@ -3,13 +3,13 @@
  * Uses the e2e-core utilities for reliable amplify init execution
  */
 
-import { AppConfiguration } from '../types';
 import { initJSProjectWithProfile } from '@aws-amplify/amplify-e2e-core';
 import { Logger } from '../utils/logger';
+import path from 'path';
+import * as fs from 'fs-extra';
 
 export interface InitializeAppOptions {
   appPath: string;
-  config: AppConfiguration;
   deploymentName: string;
   /** Amplify environment name (required, 2-10 lowercase letters) */
   envName: string;
@@ -34,7 +34,7 @@ export interface AmplifyInitSettings {
 }
 
 interface BuildInitSettingsOptions {
-  config: AppConfiguration;
+  sourceAppPath: string;
   deploymentName: string;
   envName: string;
   profile: string;
@@ -44,27 +44,16 @@ export class AmplifyInitializer {
   constructor(private readonly logger: Logger) {}
 
   async initializeApp(options: InitializeAppOptions): Promise<void> {
-    const { appPath, config, deploymentName, envName, profile } = options;
+    const { appPath, deploymentName, envName, profile } = options;
 
-    this.logger.info(`Starting amplify init for ${deploymentName} (config: ${config.app.name})`);
+    this.logger.info(`Starting amplify init for ${deploymentName}`);
     this.logger.debug(`App path: ${appPath}`);
-    this.logger.debug(`Configuration: ${JSON.stringify(config, null, 2)}`);
     this.logger.debug(`Deployment name: ${deploymentName}`);
-
-    const appNameValidation = this.validateAppName(deploymentName);
-    if (!appNameValidation.valid) {
-      throw Error(`Invalid app name: ${appNameValidation.error}`);
-    }
-
-    const amplifyEnvNameValidation = this.validateEnvName(envName);
-    if (!amplifyEnvNameValidation.valid) {
-      throw Error(`Invalid env name: ${amplifyEnvNameValidation.error}`);
-    }
 
     const startTime = Date.now();
     try {
       this.logger.debug(`Calling initJSProjectWithProfile...`);
-      const settings = this.buildInitSettings({ config, deploymentName, profile, envName });
+      const settings = this.buildInitSettings({ sourceAppPath: appPath, deploymentName, profile, envName });
       this.logger.debug(`Init settings: ${JSON.stringify(settings, null, 2)}`);
       await initJSProjectWithProfile(appPath, settings);
 
@@ -77,45 +66,6 @@ export class AmplifyInitializer {
     }
   }
 
-  validateAppName(appName: string): { valid: boolean; error?: string } {
-    // Amplify app names must be alphanumeric only, 3-20 characters
-    if (!appName) {
-      return { valid: false, error: 'App name is required' };
-    }
-
-    if (appName.length < 3 || appName.length > 20) {
-      return { valid: false, error: 'App name must be between 3-20 characters' };
-    }
-
-    // Check for alphanumeric only (no dashes, underscores, or special characters)
-    const alphanumericRegex = /^[a-zA-Z0-9]+$/;
-    if (!alphanumericRegex.test(appName)) {
-      return {
-        valid: false,
-        error: `App name not valid: ${appName}. App name must contain only alphanumeric characters (a-z, A-Z, 0-9). No dashes, underscores, or special characters allowed`,
-      };
-    }
-
-    return { valid: true };
-  }
-
-  validateEnvName(envName: string): { valid: boolean; error?: string } {
-    // Env names must be lowercase letters only, 2-10 characters
-    if (!envName) {
-      return { valid: false, error: 'Env name is required' };
-    }
-
-    const isValid = /^[a-z]{2,10}$/.test(envName);
-    if (!isValid) {
-      return {
-        valid: false,
-        error: `Env name not valid: ${envName}. Env name must be 2-10 lowercase letters only (a-z). No numbers, dashes, underscores, or special characters allowed`,
-      };
-    }
-
-    return { valid: true };
-  }
-
   /** Generates a random env name (2-10 lowercase letters) */
   static generateRandomEnvName(): string {
     const length = Math.floor(Math.random() * 9) + 2;
@@ -123,26 +73,22 @@ export class AmplifyInitializer {
   }
 
   private buildInitSettings(options: BuildInitSettingsOptions): Partial<AmplifyInitSettings> {
-    const { config, deploymentName, profile, envName } = options;
-
+    const mainTsx = path.join(options.sourceAppPath, 'src', 'main.tsx');
+    const framework = fs.existsSync(mainTsx) ? 'react' : 'none';
     const settings = {
-      name: deploymentName,
-      envName,
+      name: options.deploymentName,
+      envName: options.envName,
       editor: 'Visual Studio Code',
-      framework: config.app.framework ?? 'react',
+      framework: framework,
       srcDir: 'src',
       distDir: 'dist',
       buildCmd: 'npm run build',
       startCmd: 'npm run start',
       disableAmplifyAppCreation: false, // always create app in Amplify console
-      profileName: profile,
+      profileName: options.profile,
     };
 
-    // Log the settings being used
-    this.logger.debug(`Built init settings for ${deploymentName} (config: ${config.app.name}):`);
-    this.logger.debug(`- Name: ${settings.name}`);
-    this.logger.info(`Using Amplify environment name: ${settings.envName}`);
-    this.logger.debug(`- Using default selections for editor, framework, etc.`);
+    this.logger.debug(`Built init settings for ${options.deploymentName}: ${JSON.stringify(settings, null, 2)}`);
 
     return settings;
   }
