+++ packages/amplify-gen2-migration-e2e-system/src/core/index.ts
@@ -1,6 +1,2 @@
-export { AppSelector } from './app-selector';
-export { EnvironmentDetector } from './environment-detector';
-export { ConfigurationLoader } from './configuration-loader';
 export { AmplifyInitializer } from './amplify-initializer';
-export { CDKAtmosphereIntegration } from './cdk-atmosphere-integration';
 export { Gen2MigrationExecutor } from './gen2-migration-executor';
