+++ packages/amplify-gen2-migration-e2e-system/src/index.ts
@@ -3,15 +3,8 @@
  * Exports all public interfaces and classes
  */
 
-// Types
-export * from './types';
-
 // Core components
-export { ConfigurationLoader } from './core/configuration-loader';
-export { EnvironmentDetector } from './core/environment-detector';
-export { AppSelector } from './core/app-selector';
 export { AmplifyInitializer } from './core/amplify-initializer';
 
 // Utilities
 export { Logger } from './utils/logger';
-export { FileManager } from './utils/file-manager';
